+++ HISTORY.md
@@ -56,6 +56,7 @@
 ### 2021-11-11
 * German translation update, changed %hsva -> %clrdim
 * Brazilian Portuguese translation update, thank you, Cassiano D'Andrea!!
+* threads, gui: experimental "disalbe click-to-run" preference
 
 ### 2021-11-10
 * objects, store: new "penColorModel" setting, can e 'hsv' or 'hsl'